backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers==4.36.0
torch==2.2.0
python-multipart==0.0.6
python-docx==0.8.11
pymupdf==1.22.0
pycryptodome==3.18.0
python-jose==3.3.0
pydantic==1.10.12
